backend/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
